Use 64-bit arithmetic for tensor buffer size calculations (#3876)
* Use 64-bit arithmetic for tensor buffer size calculations

Buffer size calculations of the form dataType.getNumOfBytes() * size were
performed in 32-bit int. For large element counts these products overflow and
wrap to a small or negative value, so BaseNDManager.validateBuffer compared the
supplied buffer against the wrong expected size, and the corresponding
allocateDirect calls allocated buffers smaller than the shape requires.

Compute these byte counts in 64-bit and narrow with Math.toIntExact so a request
that does not fit in an int throws instead of silently wrapping. The same
treatment is applied to the sibling allocation paths in the api module and in
each engine, and to the (int) shape.size() narrowings in NDManager.zeros/ones
and BaseNDManager.truncatedNormal, which could truncate rather than fail.

Behaviour is unchanged for any shape that fits in an int-sized buffer.

Adds regression coverage for the overflowing element counts.

* Update to Math.multiplyExact

* Fix test and update

---------

+++ engines/tensorflow/tensorflow-engine/src/main/java/ai/djl/tensorflow/engine/TfNDArray.java
@@ -200,7 +200,8 @@ public class TfNDArray extends NativeResource<TFE_TensorHandle> implements NDArr
             JavacppUtils.setByteBuffer(getHandle(), (ByteBuffer) buffer);
             return;
         }
-        ByteBuffer bb = getManager().allocateDirect(size * type.getNumOfBytes());
+        ByteBuffer bb =
+                getManager().allocateDirect(Math.toIntExact((long) size * type.getNumOfBytes()));
         BaseNDManager.copyBuffer(buffer, bb);
         JavacppUtils.setByteBuffer(getHandle(), bb);
     }
